refactor: <navigator /> moved to components from ui
