+++ src/app/docs/[[...slug]]/page.tsx
@@ -7,7 +7,7 @@ import { mdxComponents } from "@/components/mdx-components";
 import { cache } from "react";
 import rehypePrettyCode from "rehype-pretty-code";
 import CodeBlock from "@/components/CodeBlock";
-import ComponentNavigator from "@/components/ui/Navigator";
+import ComponentNavigator from "@/components/Navigator";
 
 interface DocFrontmatter {
   title: string;
